add a simple scraper for www.marmiton.org
Given a URL to a recipe on marmiton.org recipe-scraper.py prints out the
recipe in meal master format so that it can be easily imported into
a recipe database.

The Marmiton-specific webscraping code is completely seperated from the
object representation of the recipe and its ingredients, so adding a
scraper for other websites should be relatively easy.