create framework for other recipe websites
Similar to the Exporters class to select the file format of the output,
we now have an Importers class that handles the webscrapers for the
recipe websites. Each webscraper is a class with two static methods:

  netloc(): returns the hostname of the recipe website, e.g.
            'www.marmiton.org'
  importRecipe(page): given a string containing the content of a
            website, scrapes the recipe contained therein as a
            recipe object.

For now, we still have only a webscraper class for www.marmiton.org, but
adding new ones should now be easier.

To make development of new scrapers easier, I added a hidden command
line argument -i to recipe-scraper.py to override the hostname. This is
useful, when reading a website's content from a HTML file via a file://
address.